add favorites_controller_test

+++ app/controllers/favorites_controller.rb
@@ -4,17 +4,21 @@ class FavoritesController < ApplicationController
     @favorite = Favorite.new(
       post_id: params[:post_id]
     )
-    @favorite.user_id = @current_user.id
-    if !@favorite.save
+    if !@current_user.nil?
+      @favorite.user_id = @current_user.id
+      if !@favorite.save
       redirect_to posts_path(params[:post_id])
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    @favorite = Favorite.find_by(post_id: params[:post_id], user_id: @current_user.id)
-    if !@favorite.destroy
+    if !@current_user.nil?
+      @post = Post.find(params[:post_id])
+      @favorite = Favorite.find_by(post_id: params[:post_id], user_id: @current_user.id)
+      if !@favorite.destroy
       redirect_to posts_path(params[:post_id])
+      end
     end
   end
